pyproject.toml
[project]
name = "goit-pyweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "chekh",email = "chechirkinbogd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi-limiter (>=0.1.6,<0.2.0)",
    "redis (>=5.2.1,<6.0.0)",
    "setuptools (>=75.8.2,<76.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "cloudinary (>=1.42.2,<2.0.0)",
    "sphinx (>=8.2.3,<9.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [".", "src"]
